# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: TLOZ Data API
  version: 2.0.0-draft
  description: |
    Document-first API for TLOZ Projects, Missions and Inventory.
    The v1 routes remain available temporarily as a compatibility bridge.

    Authentication is via Bearer token (agent API key) or browser session cookie.
    If both are present, the Bearer token takes priority.
  contact:
    name: TLOZ
  license:
    name: Proprietary
    identifier: LicenseRef-Zivelo-Proprietary
servers:
  - url: /api/v1
    description: Legacy v1 compatibility API
security:
  - bearerToken: []
  - sessionCookie: []
...
paths:
...
  /containers/{containerId}:
    servers:
      - url: /api/v2
    parameters:
      - name: containerId
        in: path
        required: true
        schema: { type: string, minLength: 1, maxLength: 128 }
    get:
      operationId: getContainer
      summary: Get a canonical Container
      tags: [Containers]
      responses:
        "200": { description: Container, content: { application/json: { schema: { $ref: "#/components/schemas/ContainerResponse" } } } }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      operationId: updateContainer
      summary: Update a canonical Container
      tags: [Containers]
      parameters:
        - $ref: "#/components/parameters/IfMatch"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ContainerUpdate" }
      responses:
        "200": { description: Updated Container, content: { application/json: { schema: { $ref: "#/components/schemas/ContainerResponse" } } } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { description: Revision conflict, content: { application/json: { schema: { $ref: "#/components/schemas/ContainerContentError" } } } }
        "428": { description: If-Match required, content: { application/json: { schema: { $ref: "#/components/schemas/ContainerContentError" } } } }
    delete:
      operationId: deleteContainer
      summary: Delete an empty canonical Container
      tags: [Containers]
      parameters:
        - $ref: "#/components/parameters/IfMatch"
      responses:
        "204": { description: Container deleted. }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { description: Revision conflict. }
...
  /contents/{contentId}:
    servers:
      - url: /api/v2
    parameters:
      - name: contentId
        in: path
        required: true
        schema: { type: string, minLength: 1, maxLength: 128 }
    get:
      operationId: getContent
      summary: Get canonical Content
      tags: [Content]
      responses:
        "200": { description: Content, content: { application/json: { schema: { $ref: "#/components/schemas/ContentResponse" } } } }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      operationId: updateContent
      summary: Update canonical Content
      tags: [Content]
      parameters:
        - $ref: "#/components/parameters/IfMatch"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ContentUpdate" }
      responses:
        "200": { description: Updated Content, content: { application/json: { schema: { $ref: "#/components/schemas/ContentResponse" } } } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { description: Revision conflict. }
        "428": { description: If-Match required. }
    delete:
      operationId: deleteContent
      summary: Delete canonical Content
      tags: [Content]
      parameters:
        - $ref: "#/components/parameters/IfMatch"
      responses:
        "204": { description: Content deleted. }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { description: Revision conflict. }
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: authjs.session-token
      description: Existing NextAuth browser session. Production may use the secure cookie prefix.
    bearerToken:
      type: http
      scheme: bearer
      bearerFormat: tloz_
      description: |
        New agent API keys use `tloz_` followed by 64 hex characters.
        Legacy `zaf_` keys remain valid during the v1 grace release.
        To authenticate, include the header:
        ```
        Authorization: Bearer tloz_<64-hex-chars>
        ```
        If both Bearer token and session cookie are present, the Bearer token takes priority.
  parameters:
...
    IfMatch:
      name: If-Match
      in: header
      required: true
      description: Quoted revision ETag returned by the previous read.
      schema:
        type: string
  responses:
    BadRequest:
      description: Invalid path or query input.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
    Unauthorized:
      description: An authenticated session or valid API key is required.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
    Forbidden:
      description: The authenticated actor is not permitted to perform this operation.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
    NotFound:
      description: The requested resource does not exist.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
...
  schemas:
...
    ApiError:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message, requestId]
          properties:
            code:
              type: string
              enum: [INVALID_REQUEST, UNAUTHORIZED, FORBIDDEN, NOT_FOUND, INTERNAL_ERROR, DOCUMENT_INVALID, DOCUMENT_NOT_FOUND, DOCUMENT_REVISION_CONFLICT, PRECONDITION_REQUIRED, INVALID_ATTACHMENT_MANIFEST, ATTACHMENT_STORAGE_NOT_CONFIGURED, ATTACHMENT_STORAGE_ERROR, attachment_object_missing, attachment_object_invalid, batch_superseded]
            message:
              type: string
            fields:
              type: object
              description: Field-specific guidance for INVALID_REQUEST responses.
              additionalProperties:
                type: string
            requestId:
              type: string

    # ── Users ────────────────────────────────────────────────────────
...
    ContainerDefinition:
      type: object
      additionalProperties: false
      required: [fields, views, defaultView]
      properties:
        fields:
          type: array
          items:
            type: object
            additionalProperties: true
        views:
          type: array
          items:
            type: object
            additionalProperties: true
        defaultView:
          type: string
    Container:
      type: object
      additionalProperties: false
      required: [id, publicId, presentation, title, summary, body, definition, data, revision, createdAt, updatedAt]
      properties:
        id: { type: string }
        publicId: { type: string, minLength: 1 }
        slug: { type: string }
        presentation: { type: string, minLength: 1 }
        title: { type: string, minLength: 1 }
        summary: { type: string }
        body: { type: string }
        definition:
          $ref: "#/components/schemas/ContainerDefinition"
        data:
          type: object
          additionalProperties: true
        revision: { type: integer, minimum: 1 }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    ContainerUpdate:
      type: object
      additionalProperties: false
      properties:
        slug: { type: string }
        presentation: { type: string }
        title: { type: string }
        summary: { type: string }
        body: { type: string }
        definition: { $ref: "#/components/schemas/ContainerDefinition" }
        data: { type: object, additionalProperties: true }
    ContainerResponse:
      type: object
      required: [data]
      properties:
        data: { $ref: "#/components/schemas/Container" }
...
    Content:
      type: object
      additionalProperties: false
      required: [id, publicId, containerId, presentation, title, summary, body, data, revision, createdAt, updatedAt]
      properties:
        id: { type: string }
        publicId: { type: string, minLength: 1 }
        containerId: { type: string }
        presentation: { type: string, minLength: 1 }
        title: { type: string, minLength: 1 }
        summary: { type: string }
        body: { type: string }
        data:
          type: object
          additionalProperties: true
        revision: { type: integer, minimum: 1 }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    ContentUpdate:
      type: object
      additionalProperties: false
      properties:
        presentation: { type: string }
        title: { type: string }
        summary: { type: string }
        body: { type: string }
        data: { type: object, additionalProperties: true }
    ContentResponse:
      type: object
      required: [data]
      properties:
        data: { $ref: "#/components/schemas/Content" }
...
    ContainerContentError:
      type: object
      additionalProperties: false
      required: [code, message]
      properties:
        code:
          type: string
          enum: [STORE_INVALID, STORE_NOT_FOUND, STORE_REVISION_CONFLICT, STORE_REFERENCE_INVALID, STORE_UNAVAILABLE]
        message: { type: string }
        fields:
          type: object
          additionalProperties:
            type: string

    # ── API Keys ─────────────────────────────────────────────────────
...